AttackRange v2
Movement and pathfinding might actually work now. MIGHT.
This change doesn't make it move yet but it saves the targetPosition of
the enemy unit to aim for when moving towards a unit. Since the unit
never moved yet, the code isn't foolproof, but its something. If anyone
has an idea for pathfinding, be is Dijkstras or A* [We don't have a
heuristic function for this hex grid...without a coordinate system I
can't use Euclidean or Matthattan Distance for a heuristic calculation],
let me know.

This is what I have for alpha; it doesn't take into regards for archers
and their indirect hits yet(are we doing that?)

Again, if someone wants the Edges idea for Hexes, a new class might be
wise for neighbors (in HexTile.cs), like class Connection which holds
fromHex and toHex positions as well as movementCost of the fromHex as a
edge cost.

Also, I added a HexID for each HexTile to make identifying a Hex
Position or ID better than navigating through the List map elements in
field.cs